create lib for markers handler and arm controller
